auth_service/requirements.txt
astroid==3.3.11
blinker==1.9.0
click==8.2.1
dill==0.4.0
Flask==3.1.2
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.7
packaging==25.0
parso==0.8.5
platformdirs==4.4.0
PyJWT==2.10.1
pylint==3.3.8
tomlkit==0.13.3
Werkzeug==3.1.3
